+++ src/quickjs_utils/interrupthandler.rs
@@ -26,3 +26,41 @@ unsafe extern "C" fn interrupt_handler(_rt: *mut q::JSRuntime, _opaque: *mut c_v
         }
     })
 }
+
+#[cfg(test)]
+pub mod tests {
+    use crate::esruntimebuilder::EsRuntimeBuilder;
+    use crate::esvalue::EsValueFacade;
+    use hirofa_utils::js_utils::{JsError, Script};
+    use std::cell::RefCell;
+    use std::rc::Rc;
+    use std::sync::{Arc, Mutex};
+
+    #[test]
+    fn test_interrupt_handler() {
+        let called = Arc::new(Mutex::new(RefCell::new(false)));
+        let called2 = called.clone();
+
+        let rt = EsRuntimeBuilder::new()
+            .set_interrupt_handler(move |qjs_rt| {
+                println!("ihandler called");
+                let lck = called2.lock().unwrap();
+                *lck.borrow_mut() = true;
+                false
+            })
+            .build();
+
+        match rt.eval_sync(Script::new(
+            "test_interrupt.es",
+            "for (let x = 0; x < 10000; x++) {console.log('x' + x);}",
+        )) {
+            Ok(_) => {}
+            Err(err) => {
+                panic!("err: {}", err);
+            }
+        }
+
+        let lck = called.lock().unwrap();
+        assert!(*lck.borrow());
+    }
+}